+++ leechbot/commands/__init__.py
@@ -2,7 +2,6 @@
 
 # Import submodules to register Pyrogram handlers
 from . import admin
-from . import anime
 from . import autorename
 from . import downloads
 from . import options